# docs/openapi.yml
openapi: '3.1.0'
info:
  title: Civic Influence Graph API
  version: 0.1.0
  description: |
    RESTful JSON API for the Civic Influence Graph platform.
    All responses include provenance metadata. No identity-based targeting filters
    are exposed. Authentication required only for write operations and bulk exports.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: /api/v1
    description: Default API base path

# ------------------------------------------------------------------
# Common components
# ------------------------------------------------------------------
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    apiKeyAuth:
      type: http
      scheme: bearer
      description: API key passed as Bearer token
  schemas:
    PaginationMeta:
      type: object
      properties:
        request_id: { type: string, format: uuid }
        timestamp: { type: string, format: date-time }
        data_snapshot: { type: string, nullable: true, format: date-time }
        total_count: { type: integer }
        page: { type: integer }
        page_size: { type: integer }
...
    AuditLogEntry:
      type: object
      properties:
        id: { type: string, format: uuid }
        timestamp: { type: string, format: date-time }
        natural_language_query: { type: string }
        generated_query: { type: string }
        query_params: { type: object }
        model_id: { type: string }
        model_version: { type: string }
        result_count: { type: integer }
        client_info: { type: object }
...
    BallotRace:
      type: object
      properties:
        office: { type: string }
        candidates:
          type: array
          items:
            type: object
            properties:
              entity_id: { type: string, format: uuid }
              name: { type: string }
              party: { type: string, nullable: true }
              incumbent: { type: boolean }
              summary:
                type: object
                properties:
                  total_raised: { type: number }
                  top_sectors: { type: array, items: { type: object } }
                  top_donors: { type: array, items: { type: object } }
...
paths:
...
  /ai/audit-log:
    get:
      summary: AI query audit log
      tags: [AI]
      parameters:
        - name: start_date
          in: query
          schema: { type: string, format: date-time }
        - name: end_date
          in: query
          schema: { type: string, format: date-time }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: page_size
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: Audit log entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      entries:
                        type: array
                        items: { $ref: '#/components/schemas/AuditLogEntry' }
                  meta: { $ref: '#/components/schemas/PaginationMeta' }
    post:
      summary: Create AI audit log entry
      tags: [AI]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AuditLogEntry' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/AuditLogEntry' }
                  meta: { $ref: '#/components/schemas/PaginationMeta' }
  /ballot/races:
    get:
      summary: Ballot races for an address
      tags: [Ballot]
      parameters:
        - name: address
          in: query
          required: true
          schema: { type: string, minLength: 1, maxLength: 500 }
        - name: state
          in: query
          schema: { type: string, maxLength: 2 }
      responses:
        '200':
          description: Ballot race data
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      races:
                        type: array
                        items: { $ref: '#/components/schemas/BallotRace' }
                  meta: { $ref: '#/components/schemas/PaginationMeta' }
        '400': { description: Validation error or state not found }
...